fix: move registration files to 2026 dir to restore routes
