+++ pages/register/2026/index.tsx
@@ -1,0 +1,27 @@
+import Head from 'next/head';
+import Header from '../../../components/Header/header';
+import Registration2026 from '../../../components/Form/registration2026';
+
+export default function Register2026() {
+    return (
+        <div>
+            <Head>
+                <title>Conference Registration</title>
+                <meta name="description" content="Register for the Conference" />
+            </Head>
+
+            <img
+                src="/img/illustra.png"
+                className="color-effect"
+                alt="background-illustration"
+            />
+            <Header />
+
+            <main className="container mt-20 mb-20">
+                <Registration2026 />
+            </main>
+
+            {/* Footer is rendered globally in pages/_app.tsx; avoid duplicate here */}
+        </div>
+    );
+}
